sidecar/requirements.txt
fastapi==0.115.6
lancedb==0.18.0
Pillow==11.1.0
pyarrow==18.1.0
pydantic==2.10.5
sentence_transformers==3.3.1